Clean up some int<->size_t confusion

+++ api/api.h
@@ -94,7 +94,7 @@ int X(guru64_kosherp)(int rank, const X(iodim64) *dims,
 
 FFTW_EXTERN printer *X(mkprinter_file)(FILE *f);
 
-printer *X(mkprinter_cnt)(int *cnt);
+printer *X(mkprinter_cnt)(size_t *cnt);
 printer *X(mkprinter_str)(char *s);
 
 FFTW_EXTERN planner *X(the_planner)(void);
